Move OpportunityChangeStatusForm out of components
