+++ src/apps/investments/client/opportunities/Details/OpportunityChangeStatusForm.jsx
@@ -2,15 +2,19 @@ import axios from 'axios'
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import urls from '../../lib/urls'
-import LocalHeader from './LocalHeader/LocalHeader'
+import urls from '../../../../../lib/urls'
+import LocalHeader from '../../../../../client/components/LocalHeader/LocalHeader'
 import Button from '@govuk-react/button'
 import Link from '@govuk-react/link'
 
-import { Main, FormActions, FormStateful } from '.'
+import {
+  Main,
+  FormActions,
+  FormStateful,
+} from '../../../../../client/components'
 
-import OpportunityResource from './Resource/Opportunity'
-import FieldOpportunityStatuses from './Form/elements/FieldOpportunityStatuses'
+import OpportunityResource from '../../../../../client/components/Resource/Opportunity'
+import FieldOpportunityStatuses from '../../../../../client/components/Form/elements/FieldOpportunityStatuses'
 
 const OpportunityChangeStatusForm = ({ opportunityId }) => {
   const opportunityUrl = urls.investments.opportunities.details(opportunityId)
